fix(blackbox): adding basic auth passwords for monitoring and elasticsearch (#122)
## Current Issue

As we can see in https://alertmanager.openfoodfacts.org/#/alerts:
prometheus, kibana and alertmanager (ironic) seem down. This is because
we recently added a new password on those websites, and blackbox doesn't
have access.

## Changes

This PR adds those passwords to the requests made by blackbox. It also
changes the url of elasticsearch to curl through the reverse proxy.

The same change is necessary in
[off/off-upptime/.upptimerc.yml](https://github.com/openfoodfacts/openfoodfacts-upptime/blob/master/.upptimerc.yml),
but should be much easier (see the alert manager config).

Note that GitHub believes that I have changed 300 lines of
[container-deploy.yml](https://github.com/openfoodfacts/openfoodfacts-monitoring/pull/122/files#diff-2a9064094aa5eed66f9c6c6be0b932e07de984b35eeb996f4dc95ffe47d25bd2),
but I actually just changed five lines. The changelog on VSCode shows
the true changes (line 108). This is also why CodeQL has been
retriggered. Those alerts are not relevant.

## Also done

- [x] Add GitHub secret `MONITORING_BASIC_AUTH_PASSWORD`
- [x] Add GitHub secret `ELASTICSEARCH_BASIC_AUTH_PASSWORD`

For those secrets, I reused the same naming scheme as in
[off/off-infra/.../reverse-proxy-secrets.yml](https://github.com/openfoodfacts/openfoodfacts-infrastructure/blob/develop/ansible/host_vars/monitoring-01/reverse-proxy-secrets.yml)

---------

Co-authored-by: Alex Garel <alex@garel.org>